--- app/src/main/java/com/app/bookstore/feature/dashboard/data/BookListRequest.kt
@@ -1,10 +1,0 @@
-package com.app.bookstore.feature.dashboard.data
-
-import com.google.gson.annotations.SerializedName
-
-
-/**
- * Created by Nalan Ulusoy on 13,Mart,2022
- */
-data class BookListRequest(@SerializedName("maxResults")
-                              val maxResults: Int)
