--- app/src/main/java/com/app/bookstore/feature/dashboard/data/ImageLinks.kt
@@ -1,6 +1,0 @@
-package com.app.bookstore.feature.dashboard.data
-
-import com.google.gson.annotations.SerializedName
-
-data class ImageLinks (@SerializedName( "smallThumbnail") val smallThumbnail: String? = null,
-                       @SerializedName( "thumbnail") val thumbnail: String? = null)
