--- app/src/main/java/com/app/bookstore/feature/dashboard/data/Pdf.kt
@@ -1,9 +1,0 @@
-package com.app.bookstore.feature.dashboard.data
-
-
-/**
- * Created by Nalan Ulusoy on 04,Temmuz,2022
- */
-data class Pdf(
-    val isAvailable: Boolean
-)
