+++ app/src/main/java/com/app/bookstore/feature/dashboard/data/response/Pdf.kt
@@ -1,0 +1,9 @@
+package com.app.bookstore.feature.dashboard.data.response
+
+
+/**
+ * Created by Nalan Ulusoy on 04,Temmuz,2022
+ */
+data class Pdf(
+    val isAvailable: Boolean
+)
